Improve view helper components
