+++ module/ElasticSearch/src/ElasticSearch/View/Helper/ESPerson.php
@@ -29,6 +29,11 @@ class ESPerson extends AbstractHelper
         ];
     }
 
+    public function getType(): string
+    {
+        return 'person';
+    }
+
     /**
      * @var \ElasticSearch\VuFind\RecordDriver\ESPerson
      */
@@ -41,6 +46,7 @@ class ESPerson extends AbstractHelper
 
     public function setPerson(\ElasticSearch\VuFind\RecordDriver\ESPerson $person)
     {
+        parent::setDriver($person);
         $this->person = $person;
     }
 
@@ -302,23 +308,5 @@ class ESPerson extends AbstractHelper
     public function getDetailPageLinkLabel()
     {
         return $this->resolveLabelWithDisplayName('card.knowledge.person.page.link');
-    }
-
-    /**
-     * @param string $template
-     * @param string|null $label
-     * If not null it is treated as the localization key and will be resolved before it is merged into the template.
-     * @return string
-     */
-    public function getDetailPageLink(string $template, string $label = null): string
-    {
-        $label = is_null($label)
-            ? $this->getDetailPageLinkLabel()
-            : $this->getView()->translate($label);
-
-        $segments = ['id' => $this->getPerson()->getUniqueID()];
-        $url = $this->getView()->url('page-detail-person', $segments);
-
-        return sprintf($template, $url, $label);
     }
 }
